requirements.txt
pandas
kafka-python
boto3
s3fs